+++ packages/react-ui/storybook/stories/common/Button.stories.js
@@ -33,6 +33,12 @@ const options = {
         type: 'text'
       }
     }
+  },
+  parameters: {
+    design: {
+      type: 'figma',
+      url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1303%3A26362'
+    }
   }
 };
 export default options;