+++ packages/react-ui/storybook/stories/common/Radio.stories.js
@@ -28,6 +28,12 @@ const options = {
         type: 'text'
       }
     }
+  },
+  parameters: {
+    design: {
+      type: 'figma',
+      url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1192%3A27543'
+    }
   }
 };
 export default options;